refactor(Portal): Offload common structs to ShaderUtils
ShaderFoundry, ComputePress and RenderFlow declare common structs needed
downstream by VKBuffer, various processors etc.. Currently the build
fails on windows when forward declaring.

Solution: Offload common structs to ShaderUtils.hpp
All the using declarations, vulkan specific structs, ID declarations
are moved to a new file ShaderUtils.hpp.
As this depends on vulkan headers, the standalone header makes sense
instead of GraphicsUtils.hpp
