+++ src/MayaFlux/Buffers/VKBuffer.cpp
@@ -3,7 +3,6 @@
 #include "BufferProcessingChain.hpp"
 #include "MayaFlux/Buffers/Staging/StagingUtils.hpp"
 
-#include "MayaFlux/Portal/Graphics/ShaderFoundry.hpp"
 #include "MayaFlux/Registry/BackendRegistry.hpp"
 #include "MayaFlux/Registry/Service/BufferService.hpp"
 #include "MayaFlux/Registry/Service/ComputeService.hpp"
